// src/lib/content/generatedPosts.ts
// This file is generated by scripts/generate-content.ts. Do not edit manually.
import type { Frontmatter } from "./frontmatterSchema";

export type SerializedFrontmatter = Omit<Frontmatter, "date"> & { date?: string };

export type SerializedPostEntry = {
  slug: string;
  path: string;
  sha: string;
  frontmatter: SerializedFrontmatter;
  content: string;
};

export const posts: SerializedPostEntry[] = [
  {
    "slug": "test",
    "path": "my-blog-contents/external-posts/test/index.md",
    "sha": "1a3f8431f66d4c9e6fdbd0eaec53f1b9371b80ca",
    "frontmatter": {
      "title": "test",
      "access": "public",
      "description": "テストです",
      "topics": [
        "test",
        "web"
      ],
      "date": "2026-01-14T00:00:00.000Z"
    },
    "content": "\n# こんにちは、これはテスト記事です\n\nみえてますか？\n"
  }
];

export const postsGeneratedAt = "2026-01-15T12:34:06.404Z";